+++ test/features/core-plugin/commands/config.test.js
@@ -2,24 +2,24 @@ const { of } = require('rxjs');
 const { tap, flatMap, toArray } = require('rxjs/operators');
 
 const createChaosStub = require('../../../create-chaos-stub');
-const mocks = require('../../../mocks');
+const { MockGuild, MockTextChannel, MockMessage } = require("../../../mocks/discord.mocks");
 
 describe('Feature: !config', function () {
   beforeEach(function (done) {
     this.chaos = createChaosStub();
     this.discord = this.chaos.discord;
-    this.guild = new mocks.discord.Guild({
+    this.guild = new MockGuild({
       client: this.discord,
     });
 
-    this.channel = new mocks.discord.TextChannel({
+    this.channel = new MockTextChannel({
       guild: this.guild,
       data: {
         name: 'testChannel',
       },
     });
 
-    this.message = new mocks.discord.Message({
+    this.message = new MockMessage({
       channel: this.channel,
       client: this.discord,
       data: {},